feat(module): [auth] create AuthWrapper & move auth-related code from app to wrapper
