+++ src/modules/auth/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Signin';
 export * from './Signup';
 export * from './Forgot';
+export * from './AuthWrapper';
